Continuing to make progress on figures
I did about 80% of the heavy lifting to plot the GS results. Just need to put a loop around the code I did so that a separate page can be made for each validation population. The font size of the y-axis looks goofy, so I I need to make this aesthetic change.